+++ game-dragonquest1and2hd2dremake/CHANGELOG.md
@@ -4,6 +4,11 @@
 
 - None
 
+## [0.2.1] - 2026-03-27
+
+- Fixed: Corrected Save and Config paths (Documents)
+- Changed: Config and Save modTypes will only be available if the game is on same drive as Documents (to allow hardlinks on secondary drives)
+
 ## [0.2.0] - 2026-03-19
 
 - Fixed: Pak mods now install to Paks folder directly. Load Order support is disabled for this reason